chore: extract action code for easier updates
